Show dashboard by default, disable edit when not signed in
